requirements.txt

easydict==1.10
numpy==1.23.5
tqdm==4.64.1
torchvision==0.14.1
torch==1.13.1
opencv_python==4.6.0.66
Pillow==9.3.0
tensorboardX==2.5.1
Flask-Cors==4.0.1
face-recognition==1.3.0
python-dotenv==1.0.1
flask==2.2.5